+++ src/Symfony/Component/Form/Tests/ChoiceList/Factory/CachingFactoryDecoratorTest.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Form\Tests\ChoiceList\Factory;
 
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Form\ChoiceList\ChoiceListInterface;
 use Symfony\Component\Form\ChoiceList\Factory\CachingFactoryDecorator;
 
 /**
@@ -482,6 +483,19 @@ class CachingFactoryDecoratorTest extends TestCase
 
         $this->assertSame($view1, $this->factory->createView($list, null, null, null, null, $attr1));
         $this->assertSame($view2, $this->factory->createView($list, null, null, null, null, $attr2));
+    }
+
+    /**
+     * @group legacy
+     * @expectedDeprecation Passing false as $label to Symfony\Component\Form\ChoiceList\Factory\CachingFactoryDecorator::createView is deprecated in Symfony 4.4 and will trigger a TypeError in 5.0, pass a callable that returns false instead.
+     */
+    public function testCreateViewLabelFalseDeprecation()
+    {
+        $this->factory->createView(
+            $this->createMock(ChoiceListInterface::class),
+            null, // preferred choices
+            false // label
+        );
     }
 
     public function provideSameChoices()